third_party/WebKit/LayoutTests/http/tests/websocket/connect-error-by-no-websocket-server-closeonclose-expected.txt

CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8890/' failed: Error in connection establishment: net::ERR_CONNECTION_REFUSED
Test that WebSocket fires error event if no WebSocket Server can be connected, and close is called from onclose event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test start (Port 8890)
PASS onerror was called
PASS onclose was called
PASS successfullyParsed is true

TEST COMPLETE
